+++ Areas/Identity/Pages/Account/ExternalLogin.cshtml.cs
@@ -192,7 +192,7 @@ namespace Ezequiel_Movies.Areas.Identity.Pages.Account
                             protocol: Request.Scheme);
 
                         await _emailSender.SendEmailAsync(Input.Email, "Confirm your email",
-                            $"Please confirm your account by <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>clicking here</a>.");
+                            Services.EmailTemplateService.ConfirmEmail(HtmlEncoder.Default.Encode(callbackUrl)));
 
                         // SECURITY: Sign in the user after successful account creation
                         if (_userManager.Options.SignIn.RequireConfirmedAccount)