+++ Areas/Identity/Pages/Account/Register.cshtml.cs
@@ -155,7 +155,7 @@ namespace Ezequiel_Movies.Areas.Identity.Pages.Account
                             protocol: Request.Scheme);
 
                         await _emailSender.SendEmailAsync(Input.Email, "Confirm your email",
-                            $"Please confirm your account by <a href='{HtmlEncoder.Default.Encode(callbackUrl)}'>clicking here</a>.");
+                            EmailTemplateService.ConfirmEmail(HtmlEncoder.Default.Encode(callbackUrl)));
 
                         if (_userManager.Options.SignIn.RequireConfirmedAccount)
                         {